Fix gcc build

+++ src/sdl/src/surface.cpp
@@ -131,10 +131,10 @@ auto Surface::getPixel(uint32_t x, uint32_t y) const -> Color {
   auto* pixels = static_cast<uint32_t*>(surface->pixels);
   uint32_t pixel = pixels[static_cast<size_t>(y) * static_cast<size_t>(surface->w) + static_cast<size_t>(x)];
   
-  uint8_t r = (pixel >> 24) & 0xFF;
-  uint8_t g = (pixel >> 16) & 0xFF;
-  uint8_t b = (pixel >> 8) & 0xFF;
-  uint8_t a = pixel & 0xFF;
+  uint8_t r = static_cast<uint8_t>((pixel >> 24) & 0xFF);
+  uint8_t g = static_cast<uint8_t>((pixel >> 16) & 0xFF);
+  uint8_t b = static_cast<uint8_t>((pixel >> 8) & 0xFF);
+  uint8_t a = static_cast<uint8_t>(pixel & 0xFF);
   
   return Color{r, g, b, a};
 }
